Move methods to WebexClientWrapper class
