+++ libraries/Adapters/Microsoft.Bot.Builder.Adapters.Webex/WebexAdapter.cs
@@ -15,8 +15,6 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
 {
     public class WebexAdapter : BotAdapter
     {
-        private readonly WebexAdapterOptions _config;
-
         private readonly WebexClientWrapper _webexClient;
 
         /// <summary>
@@ -25,153 +23,11 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
         /// </summary>
         /// <param name="config">An object containing API credentials, a webhook verification token and other options.</param>
         /// <param name="webexClient">A Webex API interface.</param>
-        public WebexAdapter(WebexAdapterOptions config, WebexClientWrapper webexClient)
+        public WebexAdapter(WebexClientWrapper webexClient)
         {
-            _config = config ?? throw new ArgumentNullException(nameof(config));
-
-            if (string.IsNullOrWhiteSpace(_config.AccessToken))
-            {
-                throw new Exception("AccessToken required to create controller");
-            }
-
-            if (_config.PublicAddress == null)
-            {
-                throw new Exception("PublicAddress parameter required to receive webhooks");
-            }
-
             _webexClient = webexClient ?? throw new Exception("Could not create the Webex Teams API client");
 
-            _webexClient.CreateClient(_config.AccessToken);
-        }
-
-        /// <summary>
-        /// Gets the identity of the bot.
-        /// </summary>
-        /// <value>
-        /// The identity of the bot.
-        /// </value>
-        public Person Identity { get; private set; }
-
-        /// <summary>
-        /// Load the bot's identity via the WebEx API.
-        /// MUST be called by BotBuilder bots in order to filter messages sent by the bot.
-        /// </summary>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        public async Task GetIdentityAsync(CancellationToken cancellationToken)
-        {
-            await _webexClient.GetMeAsync(cancellationToken).ContinueWith(
-                task => { Identity = task.Result; }, TaskScheduler.Current).ConfigureAwait(false);
-
-            var id = Identity.Id;
-        }
-
-        /// <summary>
-        /// Lists all webhook subscriptions currently associated with this application.
-        /// </summary>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>A list of webhook subscriptions.</returns>
-        public async Task<WebhookList> ListWebhookSubscriptionsAsync(CancellationToken cancellationToken)
-        {
-            return await _webexClient.ListWebhooksAsync(cancellationToken).ConfigureAwait(false);
-        }
-
-        /// <summary>
-        /// Clears out and resets the list of webhook subscriptions.
-        /// </summary>
-        /// <param name="webhookList">List of webhook subscriptions to be deleted.</param>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        public async Task ResetWebhookSubscriptionsAsync(WebhookList webhookList, CancellationToken cancellationToken)
-        {
-            for (var i = 0; i < webhookList.ItemCount; i++)
-            {
-                await _webexClient.DeleteWebhookAsync(webhookList.Items[i], cancellationToken).ConfigureAwait(false);
-            }
-        }
-
-        /// <summary>
-        /// Register webhook subscriptions to start receiving message events and adaptive cards events.
-        /// </summary>
-        /// <param name="webhookPath">The path of the webhook endpoint like '/api/messages'.</param>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>An array of registered <see cref="Webhook"/>.</returns>
-        public async Task<Webhook[]> RegisterWebhookSubscriptionsAsync(string webhookPath = "api/messages", CancellationToken cancellationToken = default)
-        {
-            var webHookName = string.IsNullOrWhiteSpace(_config.WebhookName) ? "Webex Firehose" : _config.WebhookName;
-            var webHookCardsName = string.IsNullOrWhiteSpace(_config.WebhookName) ? "Webex AttachmentActions" : $"{_config.WebhookName}_AttachmentActions)";
-
-            var webhookList = await ListWebhookSubscriptionsAsync(cancellationToken).ConfigureAwait(false);
-
-            string webhookId = null;
-            string webhookCardsId = null;
-
-            for (var i = 0; i < webhookList.ItemCount; i++)
-            {
-                if (webhookList.Items[i].Name == webHookName)
-                {
-                    webhookId = webhookList.Items[i].Id;
-                }
-                else if (webhookList.Items[i].Name == webHookCardsName)
-                {
-                    webhookCardsId = webhookList.Items[i].Id;
-                }
-            }
-
-            var webhookUrl = new Uri(_config.PublicAddress + webhookPath);
-
-            var webhook = await RegisterWebhookSubscriptionAsync(webhookId, webHookName, webhookUrl, cancellationToken).ConfigureAwait(false);
-            var cardsWebhook = await RegisterAdaptiveCardsWebhookSubscriptionAsync(webhookCardsId, webHookCardsName, webhookUrl, cancellationToken).ConfigureAwait(false);
-
-            return new[] { webhook, cardsWebhook };
-        }
-
-        /// <summary>
-        /// Register a webhook subscription with Webex Teams to start receiving message events.
-        /// </summary>
-        /// <param name="hookId">The id of the webhook to be registered.</param>
-        /// <param name="webHookName">The name of the webhook to be registered.</param>
-        /// <param name="webhookUrl">The Uri of the webhook.</param>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>The registered <see cref="Webhook"/>.</returns>
-        public async Task<Webhook> RegisterWebhookSubscriptionAsync(string hookId, string webHookName, Uri webhookUrl, CancellationToken cancellationToken)
-        {
-            Webhook webhook;
-
-            if (hookId != null)
-            {
-                webhook = await _webexClient.UpdateWebhookAsync(hookId, webHookName, webhookUrl, _config.Secret, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                webhook = await _webexClient.CreateWebhookAsync(webHookName, webhookUrl, EventResource.All, EventType.All, null, _config.Secret, cancellationToken).ConfigureAwait(false);
-            }
-
-            return webhook;
-        }
-
-        /// <summary>
-        /// Register a webhook subscription with Webex Teams to start receiving events related to adaptive cards.
-        /// </summary>
-        /// <param name="hookId">The id of the webhook to be registered.</param>
-        /// <param name="webHookName">The name of the webhook to be registered.</param>
-        /// <param name="webhookUrl">The Uri of the webhook.</param>
-        /// <param name="cancellationToken">A cancellation token for the task.</param>
-        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        public async Task<Webhook> RegisterAdaptiveCardsWebhookSubscriptionAsync(string hookId, string webHookName, Uri webhookUrl, CancellationToken cancellationToken)
-        {
-            Webhook webhook;
-
-            if (hookId != null)
-            {
-                webhook = await _webexClient.UpdateAdaptiveCardsWebhookAsync(hookId, webHookName, webhookUrl, _config.Secret, _config.AccessToken, cancellationToken).ConfigureAwait(false);
-            }
-            else
-            {
-                webhook = await _webexClient.CreateAdaptiveCardsWebhookAsync(webHookName, webhookUrl, EventType.All, _config.Secret, _config.AccessToken, cancellationToken).ConfigureAwait(false);
-            }
-
-            return webhook;
+            _webexClient.InitAsync().ConfigureAwait(false);
         }
 
         /// <summary>
@@ -216,7 +72,7 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
                 {
                     if (activity.Attachments[0].ContentType == "application/vnd.microsoft.card.adaptive")
                     {
-                        responseId = await _webexClient.CreateMessageWithAttachmentsAsync(personIdOrEmail, activity.Text, activity.Attachments, _config.AccessToken, cancellationToken).ConfigureAwait(false);
+                        responseId = await _webexClient.CreateMessageWithAttachmentsAsync(personIdOrEmail, activity.Text, activity.Attachments, cancellationToken).ConfigureAwait(false);
                     }
                     else
                     {
@@ -322,7 +178,7 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
                 throw new ArgumentNullException(nameof(bot));
             }
 
-            await GetIdentityAsync(cancellationToken).ConfigureAwait(false);
+            await _webexClient.GetIdentityAsync(cancellationToken).ConfigureAwait(false);
 
             WebhookEventData payload;
             string json;
@@ -333,12 +189,9 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
                 payload = JsonConvert.DeserializeObject<WebhookEventData>(json);
             }
 
-            if (!string.IsNullOrWhiteSpace(_config.Secret))
+            if (!_webexClient.ValidateSignature(request, json))
             {
-                if (!WebexHelper.ValidateSignature(_config.Secret, request, json))
-                {
-                    throw new Exception("WARNING: Webhook received message with invalid signature. Potential malicious behavior!");
-                }
+                throw new Exception("WARNING: Webhook received message with invalid signature. Potential malicious behavior!");
             }
 
             Activity activity;
@@ -347,7 +200,7 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
             {
                 var decryptedMessage = await WebexHelper.GetDecryptedMessageAsync(payload, _webexClient.GetMessageAsync, cancellationToken).ConfigureAwait(false);
 
-                activity = WebexHelper.DecryptedMessageToActivity(decryptedMessage, Identity);
+                activity = WebexHelper.DecryptedMessageToActivity(decryptedMessage, _webexClient.Identity);
             }
             else if (payload.Resource.Name == "attachmentActions" && payload.EventType == EventType.Created)
             {
@@ -357,13 +210,13 @@ namespace Microsoft.Bot.Builder.Adapters.Webex
 
                 var jsongData = JsonConvert.DeserializeObject<AttachmentActionData>(data);
 
-                var decryptedMessage = await _webexClient.GetAttachmentActionAsync(jsongData.Id, _config.AccessToken, cancellationToken).ConfigureAwait(false);
+                var decryptedMessage = await _webexClient.GetAttachmentActionAsync(jsongData.Id, cancellationToken).ConfigureAwait(false);
 
-                activity = WebexHelper.AttachmentActionToActivity(decryptedMessage, Identity);
+                activity = WebexHelper.AttachmentActionToActivity(decryptedMessage, _webexClient.Identity);
             }
             else
             {
-                activity = WebexHelper.PayloadToActivity(payload, Identity);
+                activity = WebexHelper.PayloadToActivity(payload, _webexClient.Identity);
             }
 
             using (var context = new TurnContext(this, activity))
