+++ tests/Adapters/Microsoft.Bot.Builder.Adapters.Webex.TestBot/Startup.cs
@@ -4,7 +4,6 @@
 // Generated with Bot Builder V4 SDK Template for Visual Studio EchoBot v4.3.0
 
 using System;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -24,9 +23,9 @@ namespace Microsoft.Bot.Builder.Adapters.Webex.TestBot
             Configuration = configuration;
 
             var options = new WebexAdapterOptions(configuration["AccessToken"], new Uri(configuration["PublicAddress"]), configuration["Secret"], configuration["WebhookName"]);
-            _adapter = new WebexAdapter(options, new WebexClientWrapper());
+            var wrapper = new WebexClientWrapper(options);
 
-            RegisterWebhookAsync().Wait();
+            _adapter = new WebexAdapter(wrapper);
         }
 
         public IConfiguration Configuration { get; }
@@ -66,12 +65,6 @@ namespace Microsoft.Bot.Builder.Adapters.Webex.TestBot
 
             // app.UseHttpsRedirection();
             app.UseMvc();
-        }
-
-        public async Task RegisterWebhookAsync()
-        {
-            // await _adapter.ResetWebhookSubscriptionsAsync(webhookList);
-            await _adapter.RegisterWebhookSubscriptionsAsync();
         }
     }
 }
